requirements.txt
inquirer==3.2.4
python-dotenv==1.0.1
todoist-api-python==2.1.3